+++ src/prologue/auth/users.nim
@@ -2,13 +2,13 @@ from ../core/types import SecretKey
 
 
 type
-  User* = ref object of RootObj
+  User* = object of RootObj
     username: string
     password: SecretKey
     email: string
     firstName, lastName: string
 
-  SuperUser* = ref object of User
+  SuperUser* = object of User
 
 proc initUser*(username: string, password: SecretKey, email, firstName,
     lastName = ""): User =