+++ src/prologue/core/constants.nim
@@ -1,3 +1,3 @@
 const
-  PrologueVersion* = "0.3.4" ## The version of Prologue.
+  PrologueVersion* = "0.3.6" ## The version of Prologue.
   ProloguePrefix* = "PROLOGUE_" ## The helper prefix for Prologue.