+++ src/prologue/middlewares/middlewares.nim
@@ -1,5 +1,5 @@
-import utils, cors, clickjacking, csrf
+import utils
 
-import signedcookiesession
+import signedcookiesession, cors, clickjacking, csrf, auth
 
-export utils, cors, signedcookiesession, clickjacking, csrf
+export utils, cors, signedcookiesession, clickjacking, csrf, auth