+++ changelog.md
@@ -2,6 +2,7 @@
 
 Fixes that sessionMiddleware doesn't work when user do not register session.
 Fixes HttpHeaders and adds nil check.
+Fixes cookies containing commas fail for asynchttpserver using base64 encode.
 
 ## 0.3.4